:recycle: Refactor api module, closes https://github.com/javigp2002/LazyFood/pull/84#discussion_r1448502027
La parte del API se considera del fuente y, por ello, debe estar en la misma carpeta
