+++ test/api_controller_test.ts
@@ -1,7 +1,7 @@
 import { describe,it,beforeAll, beforeEach, afterAll } from "https://deno.land/std@0.204.0/testing/bdd.ts";
 import { assert, assertInstanceOf, assertEquals } from "https://deno.land/std@0.205.0/assert/mod.ts";
-import { app } from "../api/api.ts";
-import { ApiController } from "../api/api_controller.ts";
+import { app } from "../src/api.ts";
+import { ApiController } from "../src/api_controller.ts";
 import { MyDb, crearCalendario } from "../model/bd.ts";
 import  mockito  from "npm:ts-mockito";
 import { Jugador } from "../src/jugador.ts";
